requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-cache-url==3.4.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
marshmallow==3.18.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
